+++ src/test/java/util/DateTimeUtilTest.java
@@ -1,6 +1,6 @@
 package util;
 
-import java.time.LocalDate;
+import fixture.LocalDateFixture;
 import java.time.LocalTime;
 import org.assertj.core.api.Assertions;
 import org.assertj.core.api.SoftAssertions;
@@ -59,19 +59,5 @@ class DateTimeUtilTest {
         Assertions.assertThatThrownBy(() -> {
             DateTimeUtil.isInRange(startTime, endTime, betweenTime);
         }).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    static class LocalDateFixture {
-        // DayOfWeek
-        public static LocalDate MONDAY = LocalDate.of(2025, 2, 3);
-        public static LocalDate TUESDAY = LocalDate.of(2025, 2, 4);
-        public static LocalDate WEDNESDAY = LocalDate.of(2025, 2, 5);
-        public static LocalDate THURSDAY = LocalDate.of(2025, 2, 6);
-        public static LocalDate FRIDAY = LocalDate.of(2025, 2, 7);
-        public static LocalDate SATURDAY = LocalDate.of(2025, 2, 8);
-        public static LocalDate SUNDAY = LocalDate.of(2025, 2, 9);
-
-        // Holiday
-        public static LocalDate CHRISTMAS = LocalDate.of(2025, 12, 25);
     }
 }
